requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
numpy==1.26.1
pip==22.3.1
pytz==2023.3.post1
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.8.0
wheel==0.38.4
pandas==2.1.1
scikit-learn==1.3.2
requests